pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "toolgen"
version = "0.1.0"
description = "Multi-agent tool-use conversation generator for synthetic dataset creation"
requires-python = ">=3.10"
dependencies = [
    "click>=8.1",
    "pydantic>=2.0",
    "networkx>=3.0",
    "openai>=1.0",
    "faker>=24.0",
    "mem0ai>=0.1.0",
    "numpy>=1.26",
    "tqdm>=4.66",
    "python-dotenv>=1.0",
    "tenacity>=8.2",
    "jsonlines>=4.0",
    "rich>=13.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "pytest-asyncio>=0.23",
    "pytest-mock>=3.12",
    "ruff>=0.3",
]

[project.scripts]
toolgen = "toolgen.cli:cli"

[tool.setuptools.packages.find]
where = ["."]
include = ["toolgen*"]

[tool.ruff]
line-length = 100
target-version = "py310"
